requirements.txt
astroid==2.3.3
Click==7.0
colorama==0.4.3
Flask==1.1.1
Flask-Cors==3.0.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.18
numpy==1.17.4
pandas==0.25.3
protobuf==3.11.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
